Refactoring button lists
